+++ pyfar/level/__init__.py
@@ -2,10 +2,12 @@
 
 from .standard_levels import (
     equivalent_continuous_level,
+    exposure_level,
     sliding_equivalent_continuous_level,
 )
 
 __all__ = [
     "equivalent_continuous_level",
+    "exposure_level",
     "sliding_equivalent_continuous_level",
 ]